+++ client/package.json
@@ -31,5 +31,6 @@
   },
   "devDependencies": {
     "react-router-dom": "^5.0.1"
-  }
+  },
+	"proxy": "http://localhost:5000"
 }
