+++ client/src/components/Footer.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+
 const Footer = () => {
 	return (
 		<div className="footer">
