+++ client/src/views/Home.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import '../styles/index.css'
 
 class Home extends React.Component {
 	render () {
